requirements.txt
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
Pillow==10.4.0
qrcode==7.4.2
httpx==0.27.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
pdfplumber==0.11.4
python-dotenv==1.0.1
openai>=1.40.0
# OCR (optional – requires system tesseract)
pytesseract==0.3.13
pdf2image==1.17.0
# Gmail API integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
